+++ app/src/main/java/woowacourse/shopping/presentation/productdetail/ProductDetailActivity.kt
@@ -9,8 +9,8 @@ import android.view.View
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.shopping.R
+import woowacourse.shopping.data.cart.CartRemoteDataSource
 import woowacourse.shopping.data.cart.CartRepositoryImpl
-import woowacourse.shopping.data.cart.CartServiceHelper
 import woowacourse.shopping.data.common.PreferenceUtil
 import woowacourse.shopping.data.product.ProductRemoteDataSource
 import woowacourse.shopping.data.recentproduct.RecentProductDao
@@ -47,7 +47,7 @@ class ProductDetailActivity : AppCompatActivity(), ProductDetailContract.View {
         presenter = ProductDetailPresenter(
             view = this,
             cartRepository = CartRepositoryImpl(
-                cartRemoteDataSource = CartServiceHelper(
+                cartDataSource = CartRemoteDataSource(
                     PreferenceUtil(this),
                 ),
             ),
